requirements.txt
#
#   pip install -r requirements.txt
#

Django==3.1.7
django-tinymce==3.3.0
django-allauth==0.44.0
#Pillow==8.2.0
